Prevent crash if Duo is not set-up
Added a Nonetype check and a helpful error message that should display
in the event that a returned page has no form element, as happens
when Duo has not been configured.  I recently came across this with
someone that had been using Duo for a long period, but was
still somehow being sent to the setup page in the response when using
alohomora (causing a crash).